refactor: 웹 폰트 아이콘 및 MonthPicker 컴포넌트 리팩토링 (#34)
