+++ src/pages/user/Schedule/ScheduleHeader/index.tsx
@@ -3,6 +3,7 @@ import MonthPicker from './MonthPicker';
 import { useState } from 'react';
 import MonthPickerDetail from './MonthPickerDetail';
 import { border } from '../../../../styles';
+import { MONTHS } from './constants';
 
 interface ScheduleHeaderProps {
   currentMonth: number;
@@ -22,15 +23,46 @@ const ScheduleHeader = ({
     setIsMonthPickerDetailOpen(false);
   };
 
+  const handlePrevMonth = () => {
+    const newYear = currentMonth === 1 ? currentYear - 1 : currentYear;
+    const newMonth = currentMonth === 1 ? 12 : currentMonth - 1;
+    handleYearMonthChange(newYear, newMonth);
+  };
+
+  const handlePrevYear = () => {
+    const newYear = currentYear - 1;
+    handleYearMonthChange(newYear, currentMonth);
+  };
+
+  const handleNextMonth = () => {
+    const newYear = currentMonth === 12 ? currentYear + 1 : currentYear;
+    const newMonth = currentMonth === 12 ? 1 : currentMonth + 1;
+    handleYearMonthChange(newYear, newMonth);
+  };
+
+  const handleNextYear = () => {
+    const newYear = currentYear + 1;
+    handleYearMonthChange(newYear, currentMonth);
+  };
+
   return (
     <S.Header>
-      <MonthPicker
-        currentMonth={currentMonth}
-        currentYear={currentYear}
-        handleYearMonthChange={handleYearMonthChange}
-        isMonthPickerDetailOpen={isMonthPickerDetailOpen}
-        setIsMonthPickerDetailOpen={setIsMonthPickerDetailOpen}
-      />
+      {isMonthPickerDetailOpen ? (
+        <MonthPicker
+          content={currentYear}
+          setIsMonthPickerDetailOpen={setIsMonthPickerDetailOpen}
+          onClickLeft={handlePrevYear}
+          onClickRight={handleNextYear}
+        />
+      ) : (
+        <MonthPicker
+          content={`${currentYear} ${MONTHS[currentMonth]}`}
+          setIsMonthPickerDetailOpen={setIsMonthPickerDetailOpen}
+          onClickLeft={handlePrevMonth}
+          onClickRight={handleNextMonth}
+        />
+      )}
+
       {isMonthPickerDetailOpen && (
         <MonthPickerDetail
           currentMonth={currentMonth}
